// src/app/modelos/sst/VistaSSTs.interface.ts
export interface VistaSSTs{
    departamento: string;
    titular: string;
    teldirecto: string;
    ext: string;
    extt: string;

    folio: number;
    departamento_id: string;
    destino: string;
    salidaSt: string;
    regresoSt: string;
    horaS: string;
    horaR: string;
    fechaSol: string;
    tServicio: string;
    tTransporte: string;
    infAd: string;
    nPasajeros: string;
    pasajeros: string;
}