// src/app/modelos/sst/solicitudesst.interface.ts
export interface SolicitudSTI{
    id?: string;
    folio?: string;
    departamento_id: string;
    destino: string;
    salidaSt: string;
    regresoSt: string;
    horaS: string;
    horaR: string;
    fechaSol: string;
    tServicio: string;
    tTransporte: string;
    infAd: string;
    nPasajeros: string;
    pasajeros: string;
}